+++ riscof/main.py
@@ -140,7 +140,13 @@ and DUT plugins in the config.ini file')
         if not os.path.exists(work_dir):
             logger.debug('Creating new work directory: ' + work_dir)
             os.mkdir(work_dir)
-        elif not args.testfile:
+        elif args.command=='run':
+            if args.testfile is None:
+                logger.debug('Removing old work directory: ' + work_dir)
+                shutil.rmtree(work_dir)
+                logger.debug('Creating new work directory: ' + work_dir)
+                os.mkdir(work_dir)
+        else:
             logger.debug('Removing old work directory: ' + work_dir)
             shutil.rmtree(work_dir)
             logger.debug('Creating new work directory: ' + work_dir)
@@ -197,7 +203,7 @@ and DUT plugins in the config.ini file')
         isa_file = dut.isa_spec
         platform_file = dut.platform_spec
         
-        if args.testfile:
+        if args.command=='run' and args.testfile is not None:
             isa_file = work_dir+ '/' + (isa_file.rsplit('/', 1)[1]).rsplit('.')[0] + "_checked.yaml"
             platform_file = work_dir+ '/' + (platform_file.rsplit('/', 1)[1]).rsplit('.')[0] + "_checked.yaml"
         else:
@@ -211,8 +217,8 @@ and DUT plugins in the config.ini file')
         isa_specs = utils.load_yaml(isa_file)['hart0']
         platform_specs = utils.load_yaml(platform_file)
 
-    if args.command=='gendb' or args.command=='run' or \
-            args.command=='testlist' or args.command == 'coverage' :
+    if args.command=='gendb' or args.command=='testlist' or \
+            args.command=='coverage' :
         logger.info("Generating database for suite: "+args.suite)
         work_dir = args.work_dir
         constants.suite = args.suite
@@ -223,6 +229,18 @@ and DUT plugins in the config.ini file')
         logger.info('Database File Generated: '+constants.framework_db)
         constants.env = args.env
         logger.info('Env path set to'+constants.env)
+    elif args.command=='run' :
+        if args.dbfile is None and args.testfile is None:
+            logger.info("Generating database for suite: "+args.suite)
+            work_dir = args.work_dir
+            constants.suite = args.suite
+            constants.framework_db = os.path.join(work_dir,"database.yaml")
+            logger.debug('Suite used: '+constants.suite)
+            logger.debug('ENV used: '+ args.env)
+            dbgen.generate()
+            logger.info('Database File Generated: '+constants.framework_db)
+            constants.env = args.env
+            logger.info('Env path set to'+constants.env)
 
     if args.command == 'testlist':
         test_routines.generate_test_pool(isa_specs, platform_specs, work_dir)
@@ -294,8 +312,12 @@ and DUT plugins in the config.ini file')
 
         with open(platform_file, "r") as platfile:
             pspecs = platfile.read()
+            
+        if args.dbfile is not None and args.testfile is not None:
+            logger.error("dbfile and testfile are mutually exclusive arguments")
+            raise SystemExit
         
-        cntr_args = [args.testfile,args.no_ref_run,args.no_dut_run]
+        cntr_args = [args.dbfile,args.testfile,args.no_ref_run,args.no_dut_run]
         
         report_objects = {}
         report_objects['date'] = (datetime.now(